changes to maketitle
Instead of fixing on a quote, now the second part of the title is
modular, can still insert quote using `\quote{...}` but now its made
inside the new command: `\titleelement{...}`.
This versatibility allows to insert whatever you want as the right
element of the title, could be a photo or a graph, whatever fits.